+++ app/src/domain/user/model/user.rs
@@ -5,7 +5,7 @@ use validator::ValidationError;
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 pub struct User{
-    pub id: Option<i64>,
+    pub id: Option<i32>,
     pub username: Option<String>,
     pub token: Option<String>,
     pub password: Option<String>,
